# Adds upb_strtable_tryinsert() to increase the efficiency of descriptor loading.