This PR adds a `upb_strtable_tryinsert()` function to `upb_strtable`, to avoid a double lookup in cases where we are wanting to insert but detect duplicates. The existing function `upb_strtable_insert()` requires as a precondition that the key does not exist in the table, forcing a lookup followed by insert. `upb_strtable_tryinsert()` will return a status code if the key already existed.

This also removes some deprecated functions in `upb/def.h`. I don't expect that gRPC is using these (they only use reflection for parsing and serializing JSON, not to actually inspect the schema), so I don't expect this to be an intrusive change.

This PR regains most but not all of the performance lost in https://github.com/protocolbuffers/upb/pull/416 when we load descriptors.

```
name                                      old time/op  new time/op  delta
ArenaOneAlloc                             21.4ns ± 0%  20.9ns ± 0%   -2.55%  (p=0.000 n=12+12)
ArenaInitialBlockOneAlloc                 6.28ns ± 0%  6.94ns ± 0%  +10.46%  (p=0.000 n=12+12)
LoadDescriptor_Upb                        54.8µs ± 1%  48.8µs ± 1%  -10.91%  (p=0.000 n=12+12)
LoadAdsDescriptor_Upb                     3.49ms ± 3%  3.23ms ± 0%   -7.56%  (p=0.000 n=12+11)
LoadDescriptor_Proto2                      239µs ± 0%   239µs ± 0%   +0.08%  (p=0.028 n=12+12)
LoadAdsDescriptor_Proto2                  14.7ms ± 0%  14.7ms ± 0%   +0.19%  (p=0.000 n=10+10)
Parse_Upb_FileDesc<UseArena,Copy>         11.0µs ± 0%  10.1µs ± 0%   -8.63%  (p=0.000 n=11+12)
Parse_Upb_FileDesc<UseArena,Alias>        9.37µs ± 0%  8.65µs ± 0%   -7.75%  (p=0.000 n=11+12)
Parse_Upb_FileDesc<InitBlock,Copy>        10.5µs ± 0%   9.6µs ± 0%   -8.57%  (p=0.000 n=12+12)
Parse_Upb_FileDesc<InitBlock,Alias>       8.87µs ± 0%  8.30µs ± 0%   -6.44%  (p=0.000 n=12+12)
Parse_Proto2<FileDesc,NoArena,Copy>       29.6µs ± 0%  29.4µs ± 0%   -0.63%  (p=0.000 n=12+12)
Parse_Proto2<FileDesc,UseArena,Copy>      20.9µs ± 2%  20.8µs ± 3%     ~     (p=0.266 n=12+12)
Parse_Proto2<FileDesc,InitBlock,Copy>     16.7µs ± 0%  16.7µs ± 0%     ~     (p=0.143 n=12+12)
Parse_Proto2<FileDescSV,InitBlock,Alias>  16.8µs ± 0%  16.8µs ± 0%   -0.10%  (p=0.003 n=10+10)
SerializeDescriptor_Proto2                5.33µs ± 1%  5.36µs ± 1%     ~     (p=0.219 n=12+12)
SerializeDescriptor_Upb                   12.4µs ± 0%  12.5µs ± 0%   +0.26%  (p=0.000 n=12+12)


    FILE SIZE        VM SIZE    
 --------------  -------------- 
  +7.3%    +742  +7.6%    +696    upb/table.c
    [NEW]    +743  [NEW]    +696    upb_strtable_tryinsert
    +6.1%      +7  [ = ]       0    upb_inttable_init
    -0.2%      -8  [ = ]       0    upb_inttable_compact
  +0.1%      +1  [ = ]       0    [section .strtab]
  -1.8%    -336  -2.1%    -336    upb/def.c
    -2.2%     -64  -2.3%     -64    create_msgdef
    -7.1%     -96  -7.3%     -96    create_enumdef
    -9.1%    -176  -9.3%    -176    create_fielddef
 -13.7%    -359  [ = ]       0    [Unmapped]
  +0.0%     +48  +0.3%    +360    TOTAL
```

I can't explain why some unrelated benchmarks (like arena and parse) are showing wide swings. It seems to be a benchmarking artifact, possible related to alignment.